refactor: split large codegen files (1000+ lines) into focused modules
Split 5 oversized codegen files into 11 smaller files, each under 800 lines:

- codegen_arc.cpp (1316→601): extract CoW to codegen_arc_cow.cpp (366),
  GC cycle collector to codegen_arc_gc.cpp (353)
- codegen.cpp (1297→633): extract overload resolution and user function
  calls to codegen_call_user.cpp (666)
- codegen_expr_literal.cpp (1195→516): extract valueToString and string
  conversion helpers to codegen_tostring.cpp (678)
- codegen_fn.cpp (1145→704): extract contract checking and generic type
  system to codegen_fn_generic.cpp (445)
- codegen_stmt.cpp (1110→682): extract control flow and misc statements
  to codegen_stmt_misc.cpp (430)

Also add section separator comments to codegen.hpp for navigability.

Closes #678
