+++ src/codegen_call_user.cpp
@@ -1,0 +1,666 @@
+#include "ry/codegen.hpp"
+#include "ry/diagnostic.hpp"
+#include <climits>
+
+// ===== B4: emitUserFnCall =====
+
+llvm::Function *CodeGen::resolveOverload(const std::string &callee,
+                                          const std::vector<ExprPtr> &args,
+                                          std::vector<llvm::Value*> &outArgVals) {
+    auto fit = functions_.find(callee);
+    if (fit == functions_.end()) {
+        if (native_fn_sigs_.count(callee) || native_lib_index_.count(callee)) {
+            codegenError("no matching overload for @native function '" + callee + "'");
+        } else {
+            codegenError("undefined function: " + callee +
+                " (hint: if this function is defined later in the file, "
+                "add an explicit return type to enable forward references)");
+        }
+    }
+
+    auto &overloads = fit->second;
+
+    // Identify which args are None literals
+    std::vector<bool> isNone(args.size(), false);
+    for (size_t i = 0; i < args.size(); ++i) {
+        if (isNoneLiteral(*args[i]))
+            isNone[i] = true;
+    }
+
+    // Emit non-None args to get their types
+    std::vector<llvm::Value*> emittedArgs(args.size(), nullptr);
+    for (size_t i = 0; i < args.size(); ++i) {
+        if (!isNone[i])
+            emittedArgs[i] = emitExpr(*args[i]);
+    }
+
+    struct RankedCandidate {
+        OverloadEntry *entry;
+        int exactMatches = 0;
+        int subtypeMatches = 0;
+        int wideningMatches = 0;
+        int unionMatches = 0;
+        int anyMatches = 0;
+        int defaultsUsed = 0;
+    };
+
+    auto isBetterCandidate = [](const RankedCandidate &lhs, const RankedCandidate &rhs) {
+        if (lhs.exactMatches != rhs.exactMatches)
+            return lhs.exactMatches > rhs.exactMatches;
+        if (lhs.subtypeMatches != rhs.subtypeMatches)
+            return lhs.subtypeMatches > rhs.subtypeMatches;
+        if (lhs.wideningMatches != rhs.wideningMatches)
+            return lhs.wideningMatches > rhs.wideningMatches;
+        if (lhs.unionMatches != rhs.unionMatches)
+            return lhs.unionMatches > rhs.unionMatches;
+        if (lhs.anyMatches != rhs.anyMatches)
+            return lhs.anyMatches < rhs.anyMatches;
+        if (lhs.defaultsUsed != rhs.defaultsUsed)
+            return lhs.defaultsUsed < rhs.defaultsUsed;
+        return false;
+    };
+
+    // Filter and rank candidates
+    std::vector<RankedCandidate> candidates;
+    for (auto &entry : overloads) {
+        if (args.size() < entry.minArity || args.size() > entry.paramTypes.size())
+            continue;
+        bool match = true;
+        RankedCandidate candidate{&entry, 0, 0, 0, 0, 0,
+                                  static_cast<int>(entry.paramTypes.size() - args.size())};
+        for (size_t i = 0; i < args.size(); ++i) {
+            std::string resolvedParamTypeName =
+                i < entry.paramTypeNames.size() ? resolveTypeAlias(entry.paramTypeNames[i]) : "";
+            if (isNone[i]) {
+                if (!isOptionType(entry.paramTypes[i])) { match = false; break; }
+                continue;
+            }
+
+            if (emittedArgs[i]->getType() == entry.paramTypes[i]) {
+                candidate.exactMatches++;
+                continue;
+            }
+
+            if (auto *argST = llvm::dyn_cast<llvm::StructType>(emittedArgs[i]->getType())) {
+                if (auto *paramST = llvm::dyn_cast<llvm::StructType>(entry.paramTypes[i])) {
+                    if (isSubtypeOf(argST->getName().str(), paramST->getName().str())) {
+                        candidate.subtypeMatches++;
+                        continue;
+                    }
+                }
+            }
+
+            if (isWideningConversion(emittedArgs[i], entry.paramTypes[i], resolvedParamTypeName)) {
+                candidate.wideningMatches++;
+                continue;
+            }
+
+            if (isAnyType(entry.paramTypes[i])) {
+                // Match: any type accepts all primitives; wrapping deferred to arg building
+                candidate.anyMatches++;
+            } else if (isUnionType(resolvedParamTypeName)) {
+                std::string norm = normalizeUnionType(resolvedParamTypeName);
+                auto uIt = union_type_info_.find(norm);
+                if (uIt != union_type_info_.end()) {
+                    bool found = false;
+                    for (auto *ct : uIt->second.componentTypes) {
+                        if (ct == emittedArgs[i]->getType()) { found = true; break; }
+                    }
+                    if (!found) { match = false; break; }
+                    candidate.unionMatches++;
+                } else { match = false; break; }
+            } else if (isAnyType(emittedArgs[i]->getType()) &&
+                       canAnyHoldType(entry.paramTypes[i])) {
+                // Matching a concrete parameter from an any-typed value requires runtime unwrap,
+                // so treat it with the same low specificity as an any fallback.
+                candidate.anyMatches++;
+            } else { match = false; break; }
+        }
+        if (match)
+            candidates.push_back(candidate);
+    }
+
+    if (candidates.empty())
+        codegenError("no matching overload for '" + callee + "'");
+
+    RankedCandidate *best = &candidates[0];
+    bool ambiguous = false;
+    for (size_t i = 1; i < candidates.size(); ++i) {
+        if (isBetterCandidate(candidates[i], *best)) {
+            best = &candidates[i];
+            ambiguous = false;
+        } else if (!isBetterCandidate(*best, candidates[i])) {
+            ambiguous = true;
+        }
+    }
+
+    if (ambiguous)
+        codegenError("ambiguous call to '" + callee + "'");
+
+    auto *chosen = best->entry;
+
+    // Build final arg values (fill in None args with proper Option type, wrap union args)
+    outArgVals.clear();
+    for (size_t i = 0; i < args.size(); ++i) {
+        std::string resolvedParamTypeName =
+            i < chosen->paramTypeNames.size() ? resolveTypeAlias(chosen->paramTypeNames[i]) : "";
+        if (isNone[i]) {
+            outArgVals.push_back(buildNoneValue(chosen->paramTypes[i]));
+        } else if (emittedArgs[i]->getType() != chosen->paramTypes[i] &&
+                   isWideningConversion(emittedArgs[i], chosen->paramTypes[i], resolvedParamTypeName)) {
+            outArgVals.push_back(emitWideningConversion(emittedArgs[i], chosen->paramTypes[i]));
+        } else if (auto *sliced = tryEmitSubtypeCoerce(emittedArgs[i], chosen->paramTypes[i])) {
+            outArgVals.push_back(sliced);
+        } else if (emittedArgs[i]->getType() != chosen->paramTypes[i] &&
+                   isAnyType(chosen->paramTypes[i])) {
+            outArgVals.push_back(wrapInAny(emittedArgs[i]));
+        } else if (isAnyType(emittedArgs[i]->getType()) &&
+                   emittedArgs[i]->getType() != chosen->paramTypes[i]) {
+            outArgVals.push_back(unwrapFromAny(emittedArgs[i], chosen->paramTypes[i]));
+        } else if (emittedArgs[i]->getType() != chosen->paramTypes[i] &&
+                   isUnionType(resolvedParamTypeName)) {
+            outArgVals.push_back(wrapInUnion(emittedArgs[i], resolvedParamTypeName));
+        } else {
+            outArgVals.push_back(emittedArgs[i]);
+        }
+    }
+
+    // Fill in default values for omitted parameters
+    for (size_t i = args.size(); i < chosen->paramTypes.size(); ++i) {
+        if (isNoneLiteral(*chosen->defaultValues[i])) {
+            outArgVals.push_back(buildNoneValue(chosen->paramTypes[i]));
+            continue;
+        }
+        llvm::Value *defVal = emitExpr(*chosen->defaultValues[i]);
+        std::string resolvedParamTypeName =
+            i < chosen->paramTypeNames.size() ? resolveTypeAlias(chosen->paramTypeNames[i]) : "";
+        if (defVal->getType() != chosen->paramTypes[i] &&
+            isWideningConversion(defVal, chosen->paramTypes[i], resolvedParamTypeName)) {
+            outArgVals.push_back(emitWideningConversion(defVal, chosen->paramTypes[i]));
+        } else if (auto *sliced = tryEmitSubtypeCoerce(defVal, chosen->paramTypes[i])) {
+            outArgVals.push_back(sliced);
+        } else if (defVal->getType() != chosen->paramTypes[i] &&
+                   isAnyType(chosen->paramTypes[i])) {
+            outArgVals.push_back(wrapInAny(defVal));
+        } else if (isAnyType(defVal->getType()) &&
+                   defVal->getType() != chosen->paramTypes[i]) {
+            outArgVals.push_back(unwrapFromAny(defVal, chosen->paramTypes[i]));
+        } else if (defVal->getType() != chosen->paramTypes[i] &&
+                   isUnionType(resolvedParamTypeName)) {
+            outArgVals.push_back(wrapInUnion(defVal, resolvedParamTypeName));
+        } else {
+            outArgVals.push_back(defVal);
+        }
+    }
+
+    return chosen->func;
+}
+
+llvm::Value *CodeGen::emitUserFnCall(const std::string &callee, const std::vector<ExprPtr> &args) {
+    if (deprecated_functions_.count(callee))
+        emitDeprecationWarning(callee);
+    std::vector<llvm::Value*> argVals;
+    llvm::Function *fn = resolveOverload(callee, args, argVals);
+
+    // Find the matching overload entry (single scan for constraints + result type)
+    OverloadEntry *matchedEntry = nullptr;
+    auto fit = functions_.find(callee);
+    if (fit != functions_.end()) {
+        for (auto &entry : fit->second) {
+            if (entry.func == fn) { matchedEntry = &entry; break; }
+        }
+    }
+
+    // ARC: retain arguments that are ARC-managed before passing to callee
+    for (auto *argVal : argVals)
+        tryRetainArcSource(argVal);
+
+    // Check literal/range constraints on arguments at call site
+    if (matchedEntry) {
+        for (size_t i = 0; i < matchedEntry->paramTypeNames.size() && i < argVals.size(); ++i) {
+            std::string resolvedPtype = resolveTypeAlias(matchedEntry->paramTypeNames[i]);
+            auto constraint = parseTypeConstraint(resolvedPtype);
+            if (constraint) {
+                std::string paramName = fn->getArg(i)->getName().str();
+                emitConstraintCheck(argVals[i], *constraint, paramName);
+            }
+        }
+    }
+
+    auto bindContractValue = [&](const std::string &name, llvm::Value *val,
+                                 const std::string *typeName) {
+        llvm::AllocaInst *alloca = builder_.CreateAlloca(val->getType(), nullptr, name);
+        builder_.CreateStore(val, alloca);
+        scope_stack_.back()[name] = alloca;
+        immutable_scope_stack_.back().insert(name);
+        if (!typeName) return;
+
+        std::string resolvedType = resolveTypeAlias(*typeName);
+        if (isLowLevelTypeName(resolvedType))
+            low_level_type_names_[alloca] = resolvedType;
+        if (resolvedType.size() > 9 && resolvedType.compare(0, 9, "function(") == 0)
+            fn_type_info_[alloca] = parseFnTypeAnnotation(resolvedType);
+        auto constraint = parseTypeConstraint(resolvedType);
+        if (constraint)
+            type_constraints_[alloca] = *constraint;
+        else if (isUnionType(resolvedType))
+            union_value_types_[alloca] = normalizeUnionType(resolvedType);
+    };
+
+    auto bindMockContractParams = [&]() {
+        if (!matchedEntry) return;
+        for (size_t i = 0; i < matchedEntry->paramNames.size() && i < argVals.size(); ++i) {
+            const std::string *typeName = i < matchedEntry->paramTypeNames.size()
+                ? &matchedEntry->paramTypeNames[i]
+                : nullptr;
+            bindContractValue(matchedEntry->paramNames[i], argVals[i], typeName);
+        }
+    };
+
+    auto emitMockRequireChecks = [&]() {
+        if (!matchedEntry || !matchedEntry->preconditions || matchedEntry->preconditions->empty())
+            return;
+        pushScope();
+        bindMockContractParams();
+        for (const auto &precondition : *matchedEntry->preconditions)
+            emitContractCheck("require", callee, precondition);
+        popScope();
+    };
+
+    auto emitMockEnsureChecks = [&](llvm::Value *retVal) {
+        if (!matchedEntry || !matchedEntry->postconditions || matchedEntry->postconditions->empty() ||
+            !matchedEntry->ensureBindings)
+            return;
+
+        pushScope();
+        bindMockContractParams();
+        auto &bindings = *matchedEntry->ensureBindings;
+        if (bindings.size() == 1) {
+            bindContractValue(bindings[0], retVal, nullptr);
+        } else {
+            auto *structTy = llvm::dyn_cast<llvm::StructType>(retVal->getType());
+            if (!structTy || structTy->isLiteral() || structTy->getNumElements() != bindings.size())
+                codegenError("ensure destructuring requires tuple return; binding count does not match tuple element count");
+            for (unsigned i = 0; i < bindings.size(); ++i) {
+                llvm::Value *elem = builder_.CreateExtractValue(retVal, i);
+                bindContractValue(bindings[i], elem, nullptr);
+            }
+        }
+
+        bool savedInEnsureContext = in_ensure_context_;
+        in_ensure_context_ = true;
+        for (const auto &postcondition : *matchedEntry->postconditions)
+            emitContractCheck("ensure", callee, postcondition);
+        in_ensure_context_ = savedInEnsureContext;
+        popScope();
+    };
+
+    // In test mode, inject mock dispatch only for functions targeted by mock()
+    if (test_mode_ && mocked_functions_.count(callee)) {
+        llvm::FunctionType *mockGetTy = llvm::FunctionType::get(ptrTy_, {ptrTy_}, false);
+        llvm::FunctionCallee mockGetFn = mod_->getOrInsertFunction("__ry_mock_get", mockGetTy);
+        llvm::FunctionType *mockIncTy = llvm::FunctionType::get(
+            llvm::Type::getVoidTy(*ctx_), {ptrTy_}, false);
+        llvm::FunctionCallee mockIncFn = mod_->getOrInsertFunction("__ry_mock_increment_call", mockIncTy);
+
+        auto &nameStr = mock_name_strings_[callee];
+        if (!nameStr) nameStr = cachedGlobalString(callee, ".mock." + callee);
+        llvm::Value *mockPtr = builder_.CreateCall(mockGetFn, {nameStr}, "mock_ptr");
+        llvm::Value *nullPtr = llvm::ConstantPointerNull::get(llvm::cast<llvm::PointerType>(ptrTy_));
+        llvm::Value *isMocked = builder_.CreateICmpNE(mockPtr, nullPtr, "is_mocked");
+
+        llvm::BasicBlock *mockBB = llvm::BasicBlock::Create(*ctx_, "mock_bb", fn_);
+        llvm::BasicBlock *origBB = llvm::BasicBlock::Create(*ctx_, "orig_bb", fn_);
+        llvm::BasicBlock *mergeBB = llvm::BasicBlock::Create(*ctx_, "merge_bb", fn_);
+
+        builder_.CreateCondBr(isMocked, mockBB, origBB);
+
+        // Mock path
+        builder_.SetInsertPoint(mockBB);
+        emitMockRequireChecks();
+        builder_.CreateCall(mockIncFn, {nameStr});
+        llvm::FunctionType *fnTy = fn->getFunctionType();
+        if (fn->getReturnType()->isVoidTy()) {
+            builder_.CreateCall(fnTy, mockPtr, argVals);
+            builder_.CreateBr(mergeBB);
+
+            // Original path (void case)
+            builder_.SetInsertPoint(origBB);
+            builder_.CreateCall(fn, argVals);
+            builder_.CreateBr(mergeBB);
+
+            builder_.SetInsertPoint(mergeBB);
+            return nullptr;
+        }
+
+        llvm::Value *mockResult = builder_.CreateCall(fnTy, mockPtr, argVals, "mock_result");
+        emitMockEnsureChecks(mockResult);
+        builder_.CreateBr(mergeBB);
+        llvm::BasicBlock *mockEndBB = builder_.GetInsertBlock();
+
+        // Original path
+        builder_.SetInsertPoint(origBB);
+        llvm::Value *origResult = builder_.CreateCall(fn, argVals, "orig_result");
+        builder_.CreateBr(mergeBB);
+        llvm::BasicBlock *origEndBB = builder_.GetInsertBlock();
+
+        // Merge
+        builder_.SetInsertPoint(mergeBB);
+        llvm::PHINode *phi = builder_.CreatePHI(fn->getReturnType(), 2, "call_result");
+        phi->addIncoming(mockResult, mockEndBB);
+        phi->addIncoming(origResult, origEndBB);
+        propagateReturnTypeMeta(matchedEntry, phi);
+        propagateReturnFnTypeMeta(matchedEntry, fn, phi);
+        return phi;
+    }
+
+    if (fn->getReturnType()->isVoidTy())
+        return builder_.CreateCall(fn, argVals);
+    llvm::Value *callResult = builder_.CreateCall(fn, argVals, "calltmp");
+
+    propagateReturnTypeMeta(matchedEntry, callResult);
+
+    propagateReturnFnTypeMeta(matchedEntry, fn, callResult);
+
+    return callResult;
+}
+
+void CodeGen::emitStmt(CallStmt &s) {
+    emitCoverage(s.loc);
+    if (s.callee == "describe") {
+        emitDescribeCall(s);
+        return;
+    }
+    if (s.callee == "it") {
+        emitItCall(s);
+        return;
+    }
+    if (s.callee == "mock") {
+        emitMockCall(s);
+        return;
+    }
+    if (s.callee == "fail") {
+        emitFailCall(s);
+        return;
+    }
+    auto it = builtins_.find(s.callee);
+    if (it != builtins_.end()) {
+        it->second(s.args);
+        return;
+    }
+    auto sit = struct_types_.find(s.callee);
+    if (sit != struct_types_.end()) {
+        emitStructConstructor(sit->second, s.callee, s.args);
+        return;
+    }
+    // Intercept collection operations and route through CallExpr emitter
+    if (!s.args.empty()) {
+        bool intercept = false;
+        if (auto *ve = std::get_if<VariableExpr>(&s.args[0]->data)) {
+            llvm::AllocaInst *alloca = findVar(ve->name);
+            if (alloca) {
+                bool isList = getListElementType(alloca) != nullptr;
+                bool isSet = !isList && getSetElementType(alloca) != nullptr;
+                bool isMap = !isList && !isSet && getMapKeyType(alloca) != nullptr;
+                size_t nargs = s.args.size();
+
+                if (isList &&
+                    ((s.callee == "append" && nargs == 2) ||
+                     (s.callee == "append!" && nargs == 2) ||
+                     (s.callee == "pop" && nargs == 1) ||
+                     (s.callee == "insert" && nargs == 3) ||
+                     (s.callee == "remove_at" && nargs == 2) ||
+                     (s.callee == "remove" && nargs == 2) ||
+                     (s.callee == "sort!" && (nargs == 1 || nargs == 2)) ||
+                     (s.callee == "reverse!" && nargs == 1))) {
+                    intercept = true;
+                } else if (isSet &&
+                    ((s.callee == "add" && nargs == 2) ||
+                     (s.callee == "remove" && nargs == 2) ||
+                     (s.callee == "union" && nargs == 2) ||
+                     (s.callee == "intersection" && nargs == 2) ||
+                     (s.callee == "difference" && nargs == 2) ||
+                     (s.callee == "symmetric_difference" && nargs == 2) ||
+                     (s.callee == "is_subset" && nargs == 2) ||
+                     (s.callee == "is_superset" && nargs == 2))) {
+                    intercept = true;
+                } else if (isMap &&
+                    ((s.callee == "remove" && nargs == 2) ||
+                     (s.callee == "items" && nargs == 1) ||
+                     (s.callee == "get" && (nargs == 2 || nargs == 3)))) {
+                    intercept = true;
+                }
+            }
+        }
+        if (intercept) {
+            auto ce = std::make_unique<CallExpr>();
+            ce->callee = s.callee;
+            ce->args = std::move(s.args);
+            emitExprVariant(ce);
+            return;
+        }
+    }
+    if (tryCallOperator(s.callee, s.args))
+        return;
+    // Route all remaining calls through the unified CallExpr dispatch chain.
+    // This covers @native stdlib functions, language builtins (close, range,
+    // sleep, etc.), and user-defined functions without a hardcoded whitelist.
+    auto ce = std::make_unique<CallExpr>();
+    ce->callee = s.callee;
+    ce->args = std::move(s.args);
+    emitExprVariant(ce);
+}
+
+llvm::Value *CodeGen::toBool(llvm::Value *v) {
+    if (v->getType() == i1Ty_)
+        return v;
+    if (v->getType()->isDoubleTy())
+        return builder_.CreateFCmpONE(
+            v, llvm::ConstantFP::get(f64Ty_, 0.0), "ftobool");
+    return builder_.CreateICmpNE(
+        v, llvm::ConstantInt::get(v->getType(), 0), "itobool");
+}
+
+// ===== C stdlib function helpers =====
+
+llvm::FunctionCallee CodeGen::getStdlibMalloc() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {i64Ty_}, false);
+    return mod_->getOrInsertFunction("malloc", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibRealloc() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, i64Ty_}, false);
+    return mod_->getOrInsertFunction("realloc", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibFree() {
+    auto ty = llvm::FunctionType::get(llvm::Type::getVoidTy(*ctx_), {ptrTy_}, false);
+    return mod_->getOrInsertFunction("free", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrlen() {
+    auto ty = llvm::FunctionType::get(i64Ty_, {ptrTy_}, false);
+    return mod_->getOrInsertFunction("strlen", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibMemcpy() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, ptrTy_, i64Ty_}, false);
+    return mod_->getOrInsertFunction("memcpy", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibMemmove() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, ptrTy_, i64Ty_}, false);
+    return mod_->getOrInsertFunction("memmove", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibMemset() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, i32Ty_, i64Ty_}, false);
+    return mod_->getOrInsertFunction("memset", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrcmp() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_, ptrTy_}, false);
+    return mod_->getOrInsertFunction("strcmp", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrncmp() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_, ptrTy_, i64Ty_}, false);
+    return mod_->getOrInsertFunction("strncmp", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrstr() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, ptrTy_}, false);
+    return mod_->getOrInsertFunction("strstr", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrcasestr() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, ptrTy_}, false);
+    return mod_->getOrInsertFunction("strcasestr", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrncasecmp() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_, ptrTy_, i64Ty_}, false);
+    return mod_->getOrInsertFunction("strncasecmp", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrcpy() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, ptrTy_}, false);
+    return mod_->getOrInsertFunction("strcpy", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibStrcat() {
+    auto ty = llvm::FunctionType::get(ptrTy_, {ptrTy_, ptrTy_}, false);
+    return mod_->getOrInsertFunction("strcat", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibSnprintf() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_, i64Ty_, ptrTy_}, true);
+    return mod_->getOrInsertFunction("snprintf", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibPrintf() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_}, true);
+    return mod_->getOrInsertFunction("printf", ty);
+}
+
+llvm::FunctionCallee CodeGen::getBufferedPrintf() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_}, true);
+    return mod_->getOrInsertFunction("__ry_print_printf", ty);
+}
+
+llvm::FunctionCallee CodeGen::getSprintPrintf() {
+    auto ty = llvm::FunctionType::get(i32Ty_, {ptrTy_}, true);
+    return mod_->getOrInsertFunction("__ry_sprint_printf", ty);
+}
+
+llvm::FunctionCallee CodeGen::getStdlibExit() {
+    auto ty = llvm::FunctionType::get(llvm::Type::getVoidTy(*ctx_), {i32Ty_}, false);
+    return mod_->getOrInsertFunction("exit", ty);
+}
+
+void CodeGen::emitRuntimeError(const std::string &message, const std::string &globalName,
+                                llvm::ArrayRef<llvm::Value *> extraArgs) {
+    auto fprintfTy = llvm::FunctionType::get(i32Ty_, {ptrTy_, ptrTy_}, true);
+    auto fprintfFn = mod_->getOrInsertFunction("fprintf", fprintfTy);
+#ifdef __APPLE__
+    const char *stderrName = "__stderrp";
+#else
+    const char *stderrName = "stderr";
+#endif
+    auto *stderrGlobal = mod_->getOrInsertGlobal(stderrName, ptrTy_);
+    llvm::Value *stderrVal = builder_.CreateLoad(ptrTy_, stderrGlobal, "stderr");
+    llvm::Constant *errMsg = cachedGlobalString(message, globalName);
+    llvm::SmallVector<llvm::Value *, 4> args = {stderrVal, errMsg};
+    args.append(extraArgs.begin(), extraArgs.end());
+    builder_.CreateCall(fprintfFn, args);
+    auto exitFn = getStdlibExit();
+    builder_.CreateCall(exitFn, {llvm::ConstantInt::get(i32Ty_, 1)});
+    builder_.CreateUnreachable();
+}
+
+void CodeGen::emitBoundsError(llvm::Value *index, llvm::Value *size,
+                               const std::string &fmtMsg, const std::string &globalName) {
+    emitRuntimeError(fmtMsg, globalName, {index, size});
+}
+
+llvm::Value *CodeGen::emitNegativeIndexWrap(llvm::Value *idx, llvm::Value *wrapBase,
+                                              const std::string &prefix) {
+    llvm::Value *zero = llvm::ConstantInt::get(i64Ty_, 0);
+    llvm::Value *isNeg = builder_.CreateICmpSLT(idx, zero, prefix + "_is_neg");
+    llvm::Value *wrapped = builder_.CreateAdd(idx, wrapBase, prefix + "_wrapped");
+    return builder_.CreateSelect(isNeg, wrapped, idx, prefix + "_idx");
+}
+
+void CodeGen::emitBoundsCheck(llvm::Value *&index, llvm::Value *size,
+                               const std::string &errMsg, const std::string &globalName,
+                               const std::string &bbPrefix) {
+    if (index->getType() == i1Ty_)
+        index = builder_.CreateZExt(index, i64Ty_, "idx_ext");
+
+    // Compile-time constant check with negative index wrap-around
+    if (auto *ci = llvm::dyn_cast<llvm::ConstantInt>(index)) {
+        if (auto *cs = llvm::dyn_cast<llvm::ConstantInt>(size)) {
+            int64_t idx = ci->getSExtValue();
+            int64_t sz = static_cast<int64_t>(cs->getZExtValue());
+            if (idx < 0) idx += sz;
+            if (idx < 0 || idx >= sz)
+                codegenError("index " + std::to_string(ci->getSExtValue()) +
+                             " out of bounds (size " + std::to_string(sz) + ")");
+            index = llvm::ConstantInt::get(i64Ty_, idx);
+            return;
+        }
+    }
+
+    llvm::Value *origIndex = index;
+    index = emitNegativeIndexWrap(index, size, bbPrefix);
+
+    llvm::Value *zero = llvm::ConstantInt::get(i64Ty_, 0);
+    llvm::Value *negCheck = builder_.CreateICmpSLT(
+        index, zero, bbPrefix + "_neg");
+    llvm::Value *overCheck = builder_.CreateICmpSGE(index, size, bbPrefix + "_over");
+    llvm::Value *oob = builder_.CreateOr(negCheck, overCheck, bbPrefix + "_oob");
+    llvm::BasicBlock *oobBB = llvm::BasicBlock::Create(*ctx_, bbPrefix + ".oob", fn_);
+    llvm::BasicBlock *okBB = llvm::BasicBlock::Create(*ctx_, bbPrefix + ".ok", fn_);
+    builder_.CreateCondBr(oob, oobBB, okBB);
+    builder_.SetInsertPoint(oobBB);
+    emitBoundsError(origIndex, size, errMsg, globalName);
+    builder_.SetInsertPoint(okBB);
+}
+
+llvm::Value *CodeGen::coerceToLowLevelType(llvm::Value *val, llvm::Type *targetTy,
+                                            const std::string &typeName,
+                                            const std::string &context,
+                                            const std::string &truncName) {
+    if (val->getType() == f64Ty_ && targetTy == f32Ty_)
+        return builder_.CreateFPTrunc(val, f32Ty_, truncName);
+
+    if (val->getType() == i64Ty_ && (targetTy == i8Ty_ || targetTy == i16Ty_ || targetTy == i32Ty_)) {
+        if (auto *ci = llvm::dyn_cast<llvm::ConstantInt>(val)) {
+            int64_t v = ci->getSExtValue();
+            bool isUnsigned = (!typeName.empty() && typeName[0] == 'u');
+            bool outOfRange = false;
+            if (targetTy == i8Ty_) {
+                outOfRange = isUnsigned ? (v < 0 || v > 255) : (v < INT8_MIN || v > INT8_MAX);
+            } else if (targetTy == i16Ty_) {
+                outOfRange = isUnsigned ? (v < 0 || v > (int64_t)UINT16_MAX) : (v < INT16_MIN || v > INT16_MAX);
+            } else {
+                outOfRange = isUnsigned ? (v < 0 || v > (int64_t)UINT32_MAX) : (v < INT32_MIN || v > INT32_MAX);
+            }
+            if (outOfRange)
+                codegenError(typeName + " value out of range" + context + ": " + std::to_string(v));
+        }
+        return builder_.CreateTrunc(val, targetTy, truncName);
+    }
+
+    return nullptr;
+}
+
+auto CodeGen::lookupFnTypeInfo(llvm::Value *val)
+    -> std::unordered_map<llvm::Value*, FnTypeInfo>::iterator {
+    auto it = fn_type_info_.find(val);
+    if (it == fn_type_info_.end()) {
+        if (auto *load = llvm::dyn_cast<llvm::LoadInst>(val))
+            it = fn_type_info_.find(load->getPointerOperand());
+    }
+    return it;
+}
