+++ src/codegen_tostring.cpp
@@ -1,0 +1,678 @@
+#include "ry/codegen.hpp"
+
+llvm::Value *CodeGen::valueToString(llvm::Value *val) {
+    llvm::Type *ty = val->getType();
+
+    if (ty == anyTy_)
+        return emitAnyToString(val);
+
+    // Enum value → variant name string
+    {
+        auto evIt = enum_value_types_.find(val);
+        if (evIt == enum_value_types_.end()) {
+            if (auto *load = llvm::dyn_cast<llvm::LoadInst>(val))
+                evIt = enum_value_types_.find(load->getPointerOperand());
+        }
+        if (evIt != enum_value_types_.end()) {
+            auto &einfo = enum_types_[evIt->second];
+            if (!einfo.isADT) {
+                if (einfo.hasExplicitValues) {
+                    llvm::BasicBlock *mergeBB = llvm::BasicBlock::Create(*ctx_, "vts.enum.merge", fn_);
+                    llvm::BasicBlock *defaultBB = llvm::BasicBlock::Create(*ctx_, "vts.enum.default", fn_);
+                    auto *sw = builder_.CreateSwitch(val, defaultBB, einfo.variantCount);
+                    builder_.SetInsertPoint(mergeBB);
+                    auto *namePhi = builder_.CreatePHI(ptrTy_, einfo.variantCount + 1, "vts.enum.name");
+                    for (size_t i = 0; i < einfo.variantOrder.size(); ++i) {
+                        const auto &vname = einfo.variantOrder[i];
+                        int64_t vval = einfo.variants.at(vname);
+                        llvm::BasicBlock *caseBB = llvm::BasicBlock::Create(*ctx_, "vts.enum." + vname, fn_);
+                        sw->addCase(llvm::cast<llvm::ConstantInt>(llvm::ConstantInt::get(i64Ty_, vval, true)), caseBB);
+                        builder_.SetInsertPoint(caseBB);
+                        llvm::Value *namePtr = builder_.CreateGEP(
+                            llvm::ArrayType::get(ptrTy_, einfo.variantCount),
+                            einfo.nameArray,
+                            {llvm::ConstantInt::get(i64Ty_, 0), llvm::ConstantInt::get(i64Ty_, i)},
+                            "enum_name_ptr");
+                        llvm::Value *nameStr = builder_.CreateLoad(ptrTy_, namePtr, "enum_name");
+                        namePhi->addIncoming(nameStr, caseBB);
+                        builder_.CreateBr(mergeBB);
+                    }
+                    builder_.SetInsertPoint(defaultBB);
+                    llvm::Value *unknownStr = builder_.CreateGlobalString("?", ".enum_unknown");
+                    namePhi->addIncoming(unknownStr, defaultBB);
+                    builder_.CreateBr(mergeBB);
+                    builder_.SetInsertPoint(mergeBB);
+                    return namePhi;
+                }
+                llvm::Value *namePtr = builder_.CreateGEP(
+                    llvm::ArrayType::get(ptrTy_, einfo.variantCount),
+                    einfo.nameArray,
+                    {llvm::ConstantInt::get(i64Ty_, 0), val},
+                    "enum_name_ptr");
+                return builder_.CreateLoad(ptrTy_, namePtr, "enum_name");
+            }
+
+            // ADT enum: sprint buffer + recursive valueToString
+            auto spf = getSprintPrintf();
+            emitSprintBegin();
+
+            llvm::Value *tag = builder_.CreateExtractValue(val, 0, "vts.adt.tag");
+            llvm::Value *namePtr = builder_.CreateGEP(
+                llvm::ArrayType::get(ptrTy_, einfo.variantCount),
+                einfo.nameArray,
+                {llvm::ConstantInt::get(i64Ty_, 0), tag},
+                "vts.adt.name_ptr");
+            llvm::Value *nameStr = builder_.CreateLoad(ptrTy_, namePtr, "vts.adt.name");
+
+            llvm::AllocaInst *adtAlloca = builder_.CreateAlloca(einfo.adtType, nullptr, "vts.adt.tmp");
+            builder_.CreateStore(val, adtAlloca);
+            llvm::Value *payloadPtr = builder_.CreateStructGEP(einfo.adtType, adtAlloca, 1, "vts.adt.payload");
+
+            bool anyFields = false;
+            for (auto &[vn, vf] : einfo.variantFields)
+                if (!vf.fieldTypes.empty()) { anyFields = true; break; }
+
+            if (anyFields) {
+                llvm::BasicBlock *endBB = llvm::BasicBlock::Create(*ctx_, "vts.adt.end", fn_);
+                llvm::BasicBlock *defaultBB = llvm::BasicBlock::Create(*ctx_, "vts.adt.default", fn_);
+                auto *switchInst = builder_.CreateSwitch(tag, defaultBB, einfo.variantCount);
+
+                for (auto &[vname, vtag] : einfo.variants) {
+                    llvm::BasicBlock *caseBB = llvm::BasicBlock::Create(*ctx_, "vts.adt." + vname, fn_);
+                    switchInst->addCase(
+                        llvm::cast<llvm::ConstantInt>(llvm::ConstantInt::get(i64Ty_, vtag)), caseBB);
+                    builder_.SetInsertPoint(caseBB);
+
+                    auto fit = einfo.variantFields.find(vname);
+                    if (fit != einfo.variantFields.end() && !fit->second.fieldTypes.empty()) {
+                        llvm::Constant *openFmt = cachedGlobalString("%s(", ".vts_adt_open");
+                        builder_.CreateCall(spf, {openFmt, nameStr});
+
+                        const llvm::DataLayout &dl = mod_->getDataLayout();
+                        size_t offset = 0;
+                        for (size_t fi = 0; fi < fit->second.fieldTypes.size(); ++fi) {
+                            llvm::Type *fieldTy = fit->second.fieldTypes[fi];
+                            uint64_t align = dl.getABITypeAlign(fieldTy).value();
+                            offset = (offset + align - 1) / align * align;
+                            llvm::Value *fieldPtr = builder_.CreateGEP(
+                                llvm::Type::getInt8Ty(*ctx_), payloadPtr,
+                                {llvm::ConstantInt::get(i64Ty_, offset)},
+                                "vts.adt.field." + std::to_string(fi));
+                            llvm::Value *fieldVal = builder_.CreateLoad(fieldTy, fieldPtr, "vts.adt.fval");
+
+                            // Propagate low-level type metadata for correct signedness
+                            if (fi < fit->second.fieldTypeNames.size()) {
+                                const auto &ftName = fit->second.fieldTypeNames[fi];
+                                if (isLowLevelTypeName(ftName))
+                                    low_level_type_names_[fieldVal] = ftName;
+                            }
+
+                            if (fi > 0) {
+                                llvm::Constant *commaFmt = cachedGlobalString(", ", ".vts_adt_comma");
+                                builder_.CreateCall(spf, {commaFmt});
+                            }
+
+                            llvm::Value *fieldStr = valueToString(fieldVal);
+                            llvm::Constant *sfmt = cachedGlobalString("%s", ".vts_adt_s");
+                            builder_.CreateCall(spf, {sfmt, fieldStr});
+
+                            offset += dl.getTypeAllocSize(fieldTy);
+                        }
+
+                        llvm::Constant *closeFmt = cachedGlobalString(")", ".vts_adt_close");
+                        builder_.CreateCall(spf, {closeFmt});
+                    } else {
+                        llvm::Constant *fmt = cachedGlobalString("%s", ".vts_adt_nodata");
+                        builder_.CreateCall(spf, {fmt, nameStr});
+                    }
+                    builder_.CreateBr(endBB);
+                }
+
+                builder_.SetInsertPoint(defaultBB);
+                builder_.CreateBr(endBB);
+                builder_.SetInsertPoint(endBB);
+            } else {
+                llvm::Constant *fmt = cachedGlobalString("%s", ".vts_adt_simple");
+                builder_.CreateCall(spf, {fmt, nameStr});
+            }
+
+            return emitSprintEnd("vts.adt.str");
+        }
+    }
+
+    if (ty == errorTy_) {
+        auto spf = getSprintPrintf();
+        builder_.CreateCall(getRuntimeFn("__ry_sprint_begin",
+            llvm::Type::getVoidTy(*ctx_), {}));
+        llvm::Value *msg = builder_.CreateExtractValue(val, 0, "vts_err_msg");
+        llvm::Value *code = builder_.CreateExtractValue(val, 1, "vts_err_code");
+        llvm::Constant *fmt = cachedGlobalString("Error: %s (code: %ld)", ".vts_err_fmt");
+        builder_.CreateCall(spf, {fmt, msg, code});
+        return emitSprintEnd("vts_err_str");
+    }
+
+    if (auto *structTy = llvm::dyn_cast<llvm::StructType>(ty)) {
+        for (auto &[uname, uinfo] : union_type_info_) {
+            if (uinfo.llvmType != structTy) continue;
+
+            llvm::Value *tag = builder_.CreateExtractValue(val, 0, "vts.union.tag");
+            llvm::Value *dataBytes = builder_.CreateExtractValue(val, 1, "vts.union.data");
+            auto *dataTy = uinfo.llvmType->getElementType(1);
+            llvm::AllocaInst *dataTmp = builder_.CreateAlloca(dataTy, nullptr, "vts.union.data.tmp");
+            dataTmp->setAlignment(mod_->getDataLayout().getABITypeAlign(uinfo.llvmType));
+            builder_.CreateStore(dataBytes, dataTmp);
+
+            llvm::BasicBlock *mergeBB = llvm::BasicBlock::Create(*ctx_, "vts.union.merge", fn_);
+            llvm::BasicBlock *defaultBB = llvm::BasicBlock::Create(*ctx_, "vts.union.default", fn_);
+            llvm::SwitchInst *sw = builder_.CreateSwitch(tag, defaultBB, uinfo.componentTypes.size());
+
+            builder_.SetInsertPoint(defaultBB);
+            llvm::Constant *unknownStr = cachedGlobalString("?", ".vts_union_unknown");
+            builder_.CreateBr(mergeBB);
+
+            builder_.SetInsertPoint(mergeBB);
+            auto *phi = builder_.CreatePHI(ptrTy_, uinfo.componentTypes.size() + 1, "vts.union.str");
+            phi->addIncoming(unknownStr, defaultBB);
+
+            for (size_t i = 0; i < uinfo.componentTypes.size(); ++i) {
+                llvm::BasicBlock *caseBB = llvm::BasicBlock::Create(
+                    *ctx_, "vts.union.case" + std::to_string(i), fn_);
+                sw->addCase(llvm::ConstantInt::get(
+                    llvm::cast<llvm::IntegerType>(i64Ty_), i), caseBB);
+                builder_.SetInsertPoint(caseBB);
+
+                const auto &compName = uinfo.componentNames[i];
+
+                // Reject non-stringifiable pointer-backed variants
+                if (uinfo.componentTypes[i]->isPointerTy() && compName != "str") {
+                    codegenError("cannot convert " + compName +
+                                 " variant of union to string");
+                }
+
+                llvm::Value *innerVal = builder_.CreateLoad(
+                    uinfo.componentTypes[i], dataTmp, "vts.union.inner");
+
+                // Propagate low-level type metadata for correct signedness formatting
+                if (isLowLevelTypeName(compName))
+                    low_level_type_names_[innerVal] = compName;
+
+                llvm::Value *innerStr = valueToString(innerVal);
+
+                phi->addIncoming(innerStr, builder_.GetInsertBlock());
+                builder_.CreateBr(mergeBB);
+            }
+
+            builder_.SetInsertPoint(mergeBB);
+            return phi;
+        }
+
+        if (isOptionType(ty)) {
+            auto spf = getSprintPrintf();
+            emitSprintBegin();
+
+            llvm::Value *hasValue = builder_.CreateExtractValue(val, 0, "vts.opt.has");
+            llvm::BasicBlock *someBB = llvm::BasicBlock::Create(*ctx_, "vts.opt.some", fn_);
+            llvm::BasicBlock *noneBB = llvm::BasicBlock::Create(*ctx_, "vts.opt.none", fn_);
+            llvm::BasicBlock *endBB  = llvm::BasicBlock::Create(*ctx_, "vts.opt.end", fn_);
+
+            builder_.CreateCondBr(hasValue, someBB, noneBB);
+
+            builder_.SetInsertPoint(noneBB);
+            builder_.CreateCall(spf, {cachedGlobalString("None", ".vts_none")});
+            builder_.CreateBr(endBB);
+
+            builder_.SetInsertPoint(someBB);
+            llvm::Value *innerVal = builder_.CreateExtractValue(val, 1, "vts.opt.val");
+            propagateCollectionMetadata(val, innerVal);
+            builder_.CreateCall(spf, {cachedGlobalString("Some(", ".vts_some_pre")});
+            llvm::Value *innerStr = valueToString(innerVal);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_opt_s"), innerStr});
+            builder_.CreateCall(spf, {cachedGlobalString(")", ".vts_some_post")});
+            builder_.CreateBr(endBB);
+
+            builder_.SetInsertPoint(endBB);
+            return emitSprintEnd("vts.opt.str");
+        }
+
+        if (isResultType(ty)) {
+            auto spf = getSprintPrintf();
+            emitSprintBegin();
+
+            llvm::Value *isOk = builder_.CreateExtractValue(val, 0, "vts.res.is_ok");
+            llvm::BasicBlock *okBB  = llvm::BasicBlock::Create(*ctx_, "vts.res.ok", fn_);
+            llvm::BasicBlock *errBB = llvm::BasicBlock::Create(*ctx_, "vts.res.err", fn_);
+            llvm::BasicBlock *endBB = llvm::BasicBlock::Create(*ctx_, "vts.res.end", fn_);
+
+            builder_.CreateCondBr(isOk, okBB, errBB);
+
+            builder_.SetInsertPoint(okBB);
+            llvm::Value *okVal = builder_.CreateExtractValue(val, 1, "vts.res.ok_val");
+            propagateCollectionMetadata(val, okVal);
+            builder_.CreateCall(spf, {cachedGlobalString("Ok(", ".vts_ok_pre")});
+            llvm::Value *okStr = valueToString(okVal);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_res_s"), okStr});
+            builder_.CreateCall(spf, {cachedGlobalString(")", ".vts_ok_post")});
+            builder_.CreateBr(endBB);
+
+            builder_.SetInsertPoint(errBB);
+            llvm::Value *errVal = builder_.CreateExtractValue(val, 2, "vts.res.err_val");
+            propagateCollectionMetadata(val, errVal);
+            builder_.CreateCall(spf, {cachedGlobalString("Err(", ".vts_err_pre")});
+            llvm::Value *errStr = valueToString(errVal);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_res_e"), errStr});
+            builder_.CreateCall(spf, {cachedGlobalString(")", ".vts_err_post")});
+            builder_.CreateBr(endBB);
+
+            builder_.SetInsertPoint(endBB);
+            return emitSprintEnd("vts.res.str");
+        }
+
+        std::string name = structTy->getName().str();
+        if (struct_types_.count(name))
+            return structToString(val);
+        if (isTupleStructType(structTy))
+            return tupleToString(val, structTy);
+        codegenError("cannot convert this struct type to string: " + name);
+    }
+
+    // Fixed-length array: sprint buffer + IR loop
+    if (auto *ai = llvm::dyn_cast<llvm::AllocaInst>(val)) {
+        if (auto *arrTy = llvm::dyn_cast<llvm::ArrayType>(ai->getAllocatedType())) {
+            llvm::Type *elemTy = arrTy->getElementType();
+            uint64_t arrSize = arrTy->getNumElements();
+            auto spf = getSprintPrintf();
+
+            emitSprintBegin();
+            builder_.CreateCall(spf, {cachedGlobalString("[", ".vts_arr_lb")});
+
+            llvm::BasicBlock *condBB = llvm::BasicBlock::Create(*ctx_, "vts_arr.cond", fn_);
+            llvm::BasicBlock *bodyBB = llvm::BasicBlock::Create(*ctx_, "vts_arr.body", fn_);
+            llvm::BasicBlock *endBB  = llvm::BasicBlock::Create(*ctx_, "vts_arr.end", fn_);
+
+            llvm::AllocaInst *iVar = builder_.CreateAlloca(i64Ty_, nullptr, "vts_arr_i");
+            builder_.CreateStore(llvm::ConstantInt::get(i64Ty_, 0), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(condBB);
+            llvm::Value *iVal = builder_.CreateLoad(i64Ty_, iVar, "i");
+            builder_.CreateCondBr(
+                builder_.CreateICmpSLT(iVal, llvm::ConstantInt::get(i64Ty_, arrSize)),
+                bodyBB, endBB);
+
+            builder_.SetInsertPoint(bodyBB);
+            llvm::Value *iCur = builder_.CreateLoad(i64Ty_, iVar, "i_cur");
+
+            // Comma separator if not first element
+            llvm::BasicBlock *commaBB = llvm::BasicBlock::Create(*ctx_, "vts_arr.comma", fn_);
+            llvm::BasicBlock *elemBB  = llvm::BasicBlock::Create(*ctx_, "vts_arr.elem", fn_);
+            builder_.CreateCondBr(
+                builder_.CreateICmpSGT(iCur, llvm::ConstantInt::get(i64Ty_, 0)),
+                commaBB, elemBB);
+
+            builder_.SetInsertPoint(commaBB);
+            builder_.CreateCall(spf, {cachedGlobalString(", ", ".vts_arr_comma")});
+            builder_.CreateBr(elemBB);
+
+            builder_.SetInsertPoint(elemBB);
+            llvm::Value *elemPtr = builder_.CreateGEP(
+                arrTy, ai,
+                {llvm::ConstantInt::get(i64Ty_, 0), iCur},
+                "vts_arr_ptr");
+            llvm::Value *elem = builder_.CreateLoad(elemTy, elemPtr, "vts_arr_elem");
+            llvm::Value *elemStr = valueToString(elem);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_arr_s"), elemStr});
+
+            builder_.CreateStore(
+                builder_.CreateAdd(iCur, llvm::ConstantInt::get(i64Ty_, 1)),
+                iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(endBB);
+            builder_.CreateCall(spf, {cachedGlobalString("]", ".vts_arr_rb")});
+            return emitSprintEnd("vts_arr_str");
+        }
+    }
+
+    if (ty->isPointerTy()) {
+        // Set: sprint buffer + IR loop → {elem, elem, ...}
+        if (llvm::Type *setElemTy = getSetElementType(val)) {
+            auto spf = getSprintPrintf();
+            emitSprintBegin();
+
+            auto sf = loadSetHeader(val, "vts_set");
+
+            builder_.CreateCall(spf, {cachedGlobalString("{", ".vts_set_lb")});
+
+            llvm::BasicBlock *condBB = llvm::BasicBlock::Create(*ctx_, "vts_set.cond", fn_);
+            llvm::BasicBlock *bodyBB = llvm::BasicBlock::Create(*ctx_, "vts_set.body", fn_);
+            llvm::BasicBlock *endBB  = llvm::BasicBlock::Create(*ctx_, "vts_set.end", fn_);
+
+            llvm::AllocaInst *iVar = builder_.CreateAlloca(i64Ty_, nullptr, "vts_set_i");
+            builder_.CreateStore(llvm::ConstantInt::get(i64Ty_, 0), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(condBB);
+            llvm::Value *iVal = builder_.CreateLoad(i64Ty_, iVar, "i");
+            builder_.CreateCondBr(builder_.CreateICmpSLT(iVal, sf.len), bodyBB, endBB);
+
+            builder_.SetInsertPoint(bodyBB);
+            llvm::Value *iCur = builder_.CreateLoad(i64Ty_, iVar, "i_cur");
+            llvm::BasicBlock *commaBB = llvm::BasicBlock::Create(*ctx_, "vts_set.comma", fn_);
+            llvm::BasicBlock *elemBB  = llvm::BasicBlock::Create(*ctx_, "vts_set.elem", fn_);
+            builder_.CreateCondBr(
+                builder_.CreateICmpSGT(iCur, llvm::ConstantInt::get(i64Ty_, 0)),
+                commaBB, elemBB);
+
+            builder_.SetInsertPoint(commaBB);
+            builder_.CreateCall(spf, {cachedGlobalString(", ", ".vts_set_comma")});
+            builder_.CreateBr(elemBB);
+
+            builder_.SetInsertPoint(elemBB);
+            llvm::Value *elemPtr = builder_.CreateGEP(setElemTy, sf.elems, {iCur}, "vts_set_elem_ptr");
+            llvm::Value *elem = builder_.CreateLoad(setElemTy, elemPtr, "vts_set_elem");
+            llvm::Value *elemStr = valueToString(elem);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_set_s"), elemStr});
+
+            builder_.CreateStore(
+                builder_.CreateAdd(iCur, llvm::ConstantInt::get(i64Ty_, 1)), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(endBB);
+            builder_.CreateCall(spf, {cachedGlobalString("}", ".vts_set_rb")});
+            return emitSprintEnd("vts_set_str");
+        }
+
+        // Map: sprint buffer + IR loop → {key: val, key: val, ...}
+        llvm::Type *mapKeyTy = getMapKeyType(val);
+        llvm::Type *mapValTy = getMapValueType(val);
+        if (mapKeyTy && mapValTy) {
+            auto spf = getSprintPrintf();
+            emitSprintBegin();
+
+            auto mf = loadMapHeader(val, "vts_map");
+
+            builder_.CreateCall(spf, {cachedGlobalString("{", ".vts_map_lb")});
+
+            llvm::BasicBlock *condBB = llvm::BasicBlock::Create(*ctx_, "vts_map.cond", fn_);
+            llvm::BasicBlock *bodyBB = llvm::BasicBlock::Create(*ctx_, "vts_map.body", fn_);
+            llvm::BasicBlock *endBB  = llvm::BasicBlock::Create(*ctx_, "vts_map.end", fn_);
+
+            llvm::AllocaInst *iVar = builder_.CreateAlloca(i64Ty_, nullptr, "vts_map_i");
+            builder_.CreateStore(llvm::ConstantInt::get(i64Ty_, 0), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(condBB);
+            llvm::Value *iVal = builder_.CreateLoad(i64Ty_, iVar, "i");
+            builder_.CreateCondBr(builder_.CreateICmpSLT(iVal, mf.len), bodyBB, endBB);
+
+            builder_.SetInsertPoint(bodyBB);
+            llvm::Value *iCur = builder_.CreateLoad(i64Ty_, iVar, "i_cur");
+            llvm::BasicBlock *commaBB = llvm::BasicBlock::Create(*ctx_, "vts_map.comma", fn_);
+            llvm::BasicBlock *kvBB    = llvm::BasicBlock::Create(*ctx_, "vts_map.kv", fn_);
+            builder_.CreateCondBr(
+                builder_.CreateICmpSGT(iCur, llvm::ConstantInt::get(i64Ty_, 0)),
+                commaBB, kvBB);
+
+            builder_.SetInsertPoint(commaBB);
+            builder_.CreateCall(spf, {cachedGlobalString(", ", ".vts_map_comma")});
+            builder_.CreateBr(kvBB);
+
+            builder_.SetInsertPoint(kvBB);
+            llvm::Value *keyPtr = builder_.CreateGEP(mapKeyTy, mf.keys, {iCur}, "vts_map_key_ptr");
+            llvm::Value *keyVal = builder_.CreateLoad(mapKeyTy, keyPtr, "vts_map_key");
+            llvm::Value *keyStr = valueToString(keyVal);
+            builder_.CreateCall(spf, {cachedGlobalString("%s: ", ".vts_map_kv_fmt"), keyStr});
+
+            llvm::Value *valPtr = builder_.CreateGEP(mapValTy, mf.vals, {iCur}, "vts_map_val_ptr");
+            llvm::Value *valVal = builder_.CreateLoad(mapValTy, valPtr, "vts_map_val");
+            llvm::Value *valStr = valueToString(valVal);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_map_v_s"), valStr});
+
+            builder_.CreateStore(
+                builder_.CreateAdd(iCur, llvm::ConstantInt::get(i64Ty_, 1)), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(endBB);
+            builder_.CreateCall(spf, {cachedGlobalString("}", ".vts_map_rb")});
+            return emitSprintEnd("vts_map_str");
+        }
+
+        // List: sprint buffer + IR loop → [elem, elem, ...]
+        if (llvm::Type *listElemTy = getListElementType(val)) {
+            auto spf = getSprintPrintf();
+            emitSprintBegin();
+
+            auto lf = loadListHeader(val, "vts_list");
+
+            builder_.CreateCall(spf, {cachedGlobalString("[", ".vts_list_lb")});
+
+            llvm::BasicBlock *condBB = llvm::BasicBlock::Create(*ctx_, "vts_list.cond", fn_);
+            llvm::BasicBlock *bodyBB = llvm::BasicBlock::Create(*ctx_, "vts_list.body", fn_);
+            llvm::BasicBlock *endBB  = llvm::BasicBlock::Create(*ctx_, "vts_list.end", fn_);
+
+            llvm::AllocaInst *iVar = builder_.CreateAlloca(i64Ty_, nullptr, "vts_list_i");
+            builder_.CreateStore(llvm::ConstantInt::get(i64Ty_, 0), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(condBB);
+            llvm::Value *iVal = builder_.CreateLoad(i64Ty_, iVar, "i");
+            builder_.CreateCondBr(builder_.CreateICmpSLT(iVal, lf.len), bodyBB, endBB);
+
+            builder_.SetInsertPoint(bodyBB);
+            llvm::Value *iCur = builder_.CreateLoad(i64Ty_, iVar, "i_cur");
+            llvm::BasicBlock *commaBB = llvm::BasicBlock::Create(*ctx_, "vts_list.comma", fn_);
+            llvm::BasicBlock *elemBB  = llvm::BasicBlock::Create(*ctx_, "vts_list.elem", fn_);
+            builder_.CreateCondBr(
+                builder_.CreateICmpSGT(iCur, llvm::ConstantInt::get(i64Ty_, 0)),
+                commaBB, elemBB);
+
+            builder_.SetInsertPoint(commaBB);
+            builder_.CreateCall(spf, {cachedGlobalString(", ", ".vts_list_comma")});
+            builder_.CreateBr(elemBB);
+
+            builder_.SetInsertPoint(elemBB);
+            llvm::Value *elemPtr = builder_.CreateGEP(listElemTy, lf.data, {iCur}, "vts_list_elem_ptr");
+            llvm::Value *elem = builder_.CreateLoad(listElemTy, elemPtr, "vts_list_elem");
+            llvm::Value *elemStr = valueToString(elem);
+            builder_.CreateCall(spf, {cachedGlobalString("%s", ".vts_list_s"), elemStr});
+
+            builder_.CreateStore(
+                builder_.CreateAdd(iCur, llvm::ConstantInt::get(i64Ty_, 1)), iVar);
+            builder_.CreateBr(condBB);
+
+            builder_.SetInsertPoint(endBB);
+            builder_.CreateCall(spf, {cachedGlobalString("]", ".vts_list_rb")});
+            return emitSprintEnd("vts_list_str");
+        }
+
+        if (fn_type_info_.count(val))
+            codegenError("cannot convert function to string");
+        return val; // string pointer
+    }
+    auto mallocFn = getStdlibMalloc();
+    auto snprintfFn = getStdlibSnprintf();
+
+    if (ty == i1Ty_) {
+        llvm::Constant *trueStr = cachedGlobalString("true", ".vts_true");
+        llvm::Constant *falseStr = cachedGlobalString("false", ".vts_false");
+        return builder_.CreateSelect(val, trueStr, falseStr, "vts_bool");
+    }
+    if (ty->isDoubleTy()) {
+        llvm::Value *buf = builder_.CreateCall(mallocFn, {llvm::ConstantInt::get(i64Ty_, 64)}, "vts_buf");
+        llvm::Constant *fmt = cachedGlobalString("%g", ".vts_float_fmt");
+        builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 64), fmt, val});
+        return buf;
+    }
+    // Check low-level type metadata for ambiguous LLVM types
+    std::string llName = getLowLevelTypeName(val);
+
+    if (ty == i8Ty_) {
+        llvm::Value *buf = builder_.CreateCall(mallocFn, {llvm::ConstantInt::get(i64Ty_, 32)}, "vts_buf");
+        if (llName == "i8") {
+            llvm::Constant *fmt = cachedGlobalString("%d", ".vts_i8_fmt");
+            llvm::Value *ext = builder_.CreateSExt(val, i32Ty_, "i8_ext");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, ext});
+        } else if (llName == "u8") {
+            llvm::Constant *fmt = cachedGlobalString("%u", ".vts_u8_fmt");
+            llvm::Value *ext = builder_.CreateZExt(val, i32Ty_, "u8_ext");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, ext});
+        } else {
+            // u8 (default)
+            llvm::Constant *fmt = cachedGlobalString("%u", ".vts_u8_def_fmt");
+            llvm::Value *ext = builder_.CreateZExt(val, i32Ty_, "u8_def_ext");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, ext});
+        }
+        return buf;
+    }
+    if (ty == i16Ty_) {
+        llvm::Value *buf = builder_.CreateCall(mallocFn, {llvm::ConstantInt::get(i64Ty_, 32)}, "vts_buf");
+        if (llName == "u16") {
+            llvm::Constant *fmt = cachedGlobalString("%u", ".vts_u16_fmt");
+            llvm::Value *ext = builder_.CreateZExt(val, i32Ty_, "u16_ext");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, ext});
+        } else {
+            llvm::Constant *fmt = cachedGlobalString("%d", ".vts_i16_fmt");
+            llvm::Value *ext = builder_.CreateSExt(val, i32Ty_, "i16_ext");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, ext});
+        }
+        return buf;
+    }
+    if (ty == i32Ty_) {
+        llvm::Value *buf = builder_.CreateCall(mallocFn, {llvm::ConstantInt::get(i64Ty_, 32)}, "vts_buf");
+        if (llName == "u32") {
+            llvm::Constant *fmt = cachedGlobalString("%u", ".vts_u32_fmt");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, val});
+        } else {
+            llvm::Constant *fmt = cachedGlobalString("%d", ".vts_i32_fmt");
+            builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, val});
+        }
+        return buf;
+    }
+    if (ty == f32Ty_) {
+        llvm::Value *buf = builder_.CreateCall(mallocFn, {llvm::ConstantInt::get(i64Ty_, 64)}, "vts_buf");
+        llvm::Constant *fmt = cachedGlobalString("%g", ".vts_f32_fmt");
+        llvm::Value *ext = builder_.CreateFPExt(val, f64Ty_, "f32_ext");
+        builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 64), fmt, ext});
+        return buf;
+    }
+    // default: int (i64) or i64/u64
+    llvm::Value *buf = builder_.CreateCall(mallocFn, {llvm::ConstantInt::get(i64Ty_, 32)}, "vts_buf");
+    if (llName == "u64") {
+        llvm::Constant *fmt = cachedGlobalString("%lu", ".vts_u64_fmt");
+        builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, val});
+    } else {
+        llvm::Constant *fmt = cachedGlobalString("%ld", ".vts_int_fmt");
+        builder_.CreateCall(snprintfFn, {buf, llvm::ConstantInt::get(i64Ty_, 32), fmt, val});
+    }
+    return buf;
+}
+
+llvm::Value *CodeGen::structToString(llvm::Value *val) {
+    auto *structTy = llvm::cast<llvm::StructType>(val->getType());
+    std::string typeName = structTy->getName().str();
+    auto it = struct_types_.find(typeName);
+    if (it == struct_types_.end())
+        codegenError("structToString: unknown record type: " + typeName);
+
+    const auto &info = it->second;
+
+    // Check for user-defined to_str overload
+    auto fit = functions_.find("to_str");
+    if (fit != functions_.end()) {
+        for (auto &entry : fit->second) {
+            if (entry.paramTypes.size() == 1 && entry.paramTypes[0] == structTy) {
+                return builder_.CreateCall(entry.func, {val}, "user_to_str");
+            }
+        }
+    }
+
+    // Auto-generate: "TypeName(field1: val1, field2: val2)"
+    auto strlenFn = getStdlibStrlen();
+
+    using SP = std::pair<llvm::Value*, llvm::Value*>;
+    std::vector<SP> parts;
+
+    auto addLiteral = [&](const std::string &s, const char *label) {
+        parts.push_back({cachedGlobalString(s, label),
+                         llvm::ConstantInt::get(i64Ty_, s.size())});
+    };
+
+    addLiteral(typeName + "(", ".sts_prefix");
+
+    for (unsigned i = 0; i < info.fields.size(); ++i) {
+        if (i > 0)
+            addLiteral(", ", ".sts_sep");
+        addLiteral(info.fields[i].name + ": ", ".sts_fname");
+        llvm::Value *field = builder_.CreateExtractValue(val, i, info.fields[i].name);
+        llvm::Value *fieldStr = valueToString(field);
+        parts.push_back({fieldStr, builder_.CreateCall(strlenFn, {fieldStr}, "sts_len")});
+    }
+
+    addLiteral(")", ".sts_suffix");
+
+    return concatStringParts(parts, "sts");
+}
+
+llvm::Value *CodeGen::tupleToString(llvm::Value *val, llvm::StructType *st) {
+    auto strlenFn = getStdlibStrlen();
+
+    using SP = std::pair<llvm::Value*, llvm::Value*>;
+    std::vector<SP> parts;
+
+    auto addLiteral = [&](const std::string &s, const char *label) {
+        parts.push_back({cachedGlobalString(s, label),
+                         llvm::ConstantInt::get(i64Ty_, s.size())});
+    };
+
+    unsigned n = st->getNumElements();
+    addLiteral("(", ".tts_prefix");
+    for (unsigned i = 0; i < n; ++i) {
+        if (i > 0)
+            addLiteral(", ", ".tts_sep");
+        llvm::Value *elem = builder_.CreateExtractValue(val, i, "tts_elem");
+        llvm::Value *elemStr = valueToString(elem);
+        parts.push_back({elemStr, builder_.CreateCall(strlenFn, {elemStr}, "tts_len")});
+    }
+    if (n == 1)
+        addLiteral(",", ".tts_trail");
+    addLiteral(")", ".tts_suffix");
+
+    return concatStringParts(parts, "tts");
+}
+
+void CodeGen::emitSprintBegin() {
+    builder_.CreateCall(getRuntimeFn("__ry_sprint_begin",
+        llvm::Type::getVoidTy(*ctx_), {}));
+}
+
+llvm::Value *CodeGen::emitSprintEnd(const llvm::Twine &name) {
+    return builder_.CreateCall(
+        getRuntimeFn("__ry_sprint_end", ptrTy_, {}), {}, name);
+}
+
+llvm::Value *CodeGen::concatStringParts(
+    const std::vector<std::pair<llvm::Value*, llvm::Value*>> &parts,
+    const std::string &prefix) {
+    auto mallocFn = getStdlibMalloc();
+    auto memcpyFn = getStdlibMemcpy();
+
+    llvm::Value *totalLen = llvm::ConstantInt::get(i64Ty_, 0);
+    for (auto &p : parts)
+        totalLen = builder_.CreateAdd(totalLen, p.second, prefix + "_total");
+
+    llvm::Value *bufSize = builder_.CreateAdd(
+        totalLen, llvm::ConstantInt::get(i64Ty_, 1), prefix + "_bufsize");
+    llvm::Value *buf = builder_.CreateCall(mallocFn, {bufSize}, prefix + "_buf");
+    llvm::Value *off = llvm::ConstantInt::get(i64Ty_, 0);
+    for (auto &p : parts) {
+        llvm::Value *dst = builder_.CreateGEP(
+            builder_.getInt8Ty(), buf, {off}, prefix + "_dst");
+        builder_.CreateCall(memcpyFn, {dst, p.first, p.second});
+        off = builder_.CreateAdd(off, p.second, prefix + "_off");
+    }
+
+    llvm::Value *ep = builder_.CreateGEP(
+        builder_.getInt8Ty(), buf, {off}, prefix + "_end");
+    builder_.CreateStore(llvm::ConstantInt::get(builder_.getInt8Ty(), 0), ep);
+    return buf;
+}
